Create individual file for `getSubtitles` function
